Fix writing multiple configuration words on PIC
The items value was incorrect for writing multiple
configuration words on PIC. With the previous code the
`items` variable was always set to 1 which would have
the effect that only the first configuration word
woudl be written or read. Setting it to to `num_fuses`
is correct but we conditionalize this to fuses types
where rev_mask == 0x5 (PIC MCUs).