Makefile: don't make objects depend on version_header
The version_header target doesn't produce a file and thus is always
remade; effectively cancelling the benefits of using Make to build
things as opposed to just a shell scripts:

  $ make xml.o
  gcc -g -O0 -Wall ... -c -o xml.o xml.c
  $ make xml.o
  gcc -g -O0 -Wall ... -c -o xml.o xml.c
  $

Remove it from the dependency chain. We depend on the actual file
already. Actually -- we depend on it multiple times. Let's remove the
redundant dependences, it seems to improve the clarity.