dumpic/dump-lgc: Add a tool to produce XML dumps from .LGC files
The LGC files are exported by the XGecu programmer tool for Windows.

This tools produces a file that could eventually be used by the minipro
tool (and possibly other logic chip testing tools out there, as there's
nothing specific to Minipro hardware or software).

The structure of the output file is the same as the infoic.xml with the
<ic> element being somewhat different:

  <ic name="7400" type="5" voltage="5V" pins="14">
    <vector id="00"> 0 0 H 0 0 H V H 0 0 H 0 0 G </vector>
    <vector id="01"> 1 0 H 1 0 H V H 1 0 H 1 0 G </vector>
    <vector id="02"> 0 1 H 0 1 H V H 0 1 H 0 1 G </vector>
    <vector id="03"> 1 1 L 1 1 L V L 1 1 L 1 1 G </vector>
  </ic>

The vector element body defines pin states with plain letters
(whitespace ignored) as opposed to XML elements for the sake of clarity.
XML really doesn't shine here, just ponder this:

  <vector id="00"><pin state="1" /><pin state="Z" />...</vector>

or perhaps:

  <vector id="00"><output state="1" /><tristate />...</vector>

With our SAX parser's lack of lookahead this would be a nightmare to
parse as well.

I've considered writing it in Python for consistency with the other tool in
dumpic, but given how inferior its unpack is it would be considerably uglier
and more difficult than doing it in Perl. Hope that's okay.