Write the full JEDEC row when storing a GAL UES
This commit is in preparation for adding support for the GAL20RA10B. For this
chip, if writing less than the full JEDEC row when storing the User Electronic
Signature (UES) the data is not stored correctly - it is shifted by the
difference between the row_width (80) and ues_size (64) resulting in a wrong
UES and verification failure from minipro. This commit will allow successful
programming and verification of the UES for the GAL20RA10B via configuration to
be provided in a separate commit.

I have tested other chips and confirmed they still program and verify
successfully after this commit:
  Atmel   ATF22V10C  - using the ATF22V10C(UES) programming option
  Lattice GAL20V8B

I believe this commit introduces 'correct' UES programming behavior for GALs in
general. At least some other GALs (ATF22V10C and GAL20V8B) are tolerant of the
UES being written with less than the full number of bits for the JEDEC row, but
will still program correctly when provided with extra padding bits to the
lenghth of the full JEDEC row as described in this link:
(http://www.armory.com/~rstevew/Public/Pgmrs/GAL/algo.htm).

The linked document describes fuse maps for various GAL types, and shows
padding bits after the UES. For example for the GAL22V10 it shows padding of 68
'low' bits following the 64 UES bits, taking the row to the length of 132
matching the other rows in the fuse map:

  Fusemap: RA=0:  0, 44,  88, ... 5764, L, L, L, L, L, L (132+6   bit, line  0)
           RA=0:  1, 45,  89, ... 5765, H, L, L, L, L, L (132+6   bit, line  1)
           :
           RA=0: 43, 87, 131, ... 5807, H, H, L, H, L, H (132+6   bit, line 43)
  UES:     RA=0: 5828-5891,       68*L, L, L, H, H, L, H (64+68+6 bit, line 44)