Add support for Xicor X28HC256
This EEPROM is mostly identical to the Xicor X28C256 (most of the
datasheet is word-for-word identical), except that the HC has 128-byte
pages and the C has 64-byte pages (plus some timing differences).

This copies the X28C256 config and modifies just the page size. Has been
tested with a TL866CS and a X28HC256-15 EEPROM.