Add support for Hitachi HN43128 mask programmable ROM. This has been tested in Fedora Linux, using a TL866II Plus.
To facilitate mask programmable CS and CE, which may be programmed to either be active high or active low, a new
mode was introduced in prom.c, to loop over all possible combinations of values in sequence, until a non-empty
sequence is found. If none is found, ROM is assumed to contain all 0xFFs.

This can be optimised - once a valid set of CS and CE pins is found, this can be cached to device. This has not
been done.

Furthermore, a HN43128 can be programmed either as an 8 bit device, or as a 4 bit device with twice the number
of words. The 4 bit mode has not been handled, however doing so should be trivial.

This code can easily be extended to cover other Hitachi mask PROMs that work with a similar scheme, such as the
HN61256. Doing so was not done here, due to lack of devices to test.